+++ zk_ip/guest/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2021"
 [dependencies]
 airbender = { package = "airbender-sdk", git = "https://github.com/popzxc/airbender-platform", branch = "main", features = ["std", "crypto"] }
 alloy-sol-types = { version = "1.5.7", default-features = false }
-
+utils = { package = "zkip-host", path = "../host" }
